backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.7.3
python-dotenv==1.0.1
bcrypt==4.1.3
cloudinary==1.40.0
Pillow==10.4.0
scikit-image==0.23.2
opencv-python-headless==4.10.0.84
numpy==1.26.4
requests==2.32.3
openpyxl==3.1.4
reportlab==4.2.2
gunicorn==22.0.0
